Cambiar unas pocas cosas

+++ demo/src/main/java/com/example/demo/repository/HabitacionRepositorio.java
@@ -3,7 +3,7 @@ package com.example.demo.repository;
 import com.example.demo.model.Habitacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HabitacionRepositorio extends JpaRepository<Habitacion, Long> {
+public interface HabitacionRepositorio extends JpaRepository<Habitacion, Integer> {
 
 
 }
